requirements.txt
black==24.1.1
bootstrap4==0.1.0
colorama==0.4.6
crispy-bootstrap4==2023.1
Django==5.0.2
django-cleanup==8.1.0
django-crispy-forms==2.1
isort==5.13.2
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-oauthlib==1.2.0
matplotlib==3.8.3
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
python-dotenv==1.0.1
requests==2.31.0
scipy==1.12.0
social-auth-app-django==5.4.0
social-auth-core==4.4.2
tensorflow==2.15.0
typing_extensions==4.9.0
